+++ Westwind.Utilities.Test/ReflectionUtilsTests.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using System.Reflection;
 
 namespace Westwind.Utilities.Tests
 {
@@ -39,5 +40,50 @@ namespace Westwind.Utilities.Tests
             val2 = ReflectionUtils.StringToTypedValue<Single>(res);
             Assert.AreEqual(val, val2);
         }
+
+        //[TestMethod]
+        //public void ComAccessReflectionCoreAnd45Test()
+        //{
+        //    // this works with both .NET 4.5+ and .NET Core 2.0+
+
+        //    string progId = "InternetExplorer.Application";
+        //    Type type = Type.GetTypeFromProgID(progId);
+        //    object inst = Activator.CreateInstance(type);
+
+
+        //    inst.GetType().InvokeMember("Visible", ReflectionUtils.MemberAccess | BindingFlags.SetProperty, null, inst,
+        //        new object[1]
+        //        {
+        //            true
+        //        });
+
+        //    inst.GetType().InvokeMember("Navigate", ReflectionUtils.MemberAccess | BindingFlags.InvokeMethod, null,
+        //        inst, new object[]
+        //        {
+        //            "https://markdownmonster.west-wind.com",
+        //        });
+
+        //    //result = ReflectionUtils.GetPropertyCom(inst, "cAppStartPath");
+        //    bool result = (bool)inst.GetType().InvokeMember("Visible",
+        //        ReflectionUtils.MemberAccess | BindingFlags.GetProperty, null, inst, null);
+        //    Console.WriteLine(result); // path             
+        //}
+
+        //[TestMethod]
+        //public void ComAccessDynamicCoreAnd45Test()
+        //{
+        //    // this does not work with .NET Core 2.0
+
+        //    string progId = "InternetExplorer.Application";
+        //    Type type = Type.GetTypeFromProgID(progId);
+
+        //    dynamic inst = Activator.CreateInstance(type);
+
+        //    inst.Visible = true;
+        //    inst.Navigate("https://markdownmonster.west-wind.com");
+
+        //    bool result = inst.Visible;
+        //    Assert.IsTrue(result);
+        //}
     }
 }