+++ Westwind.Utilities/Utilities/DebugUtils.cs
@@ -33,6 +33,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.Text;
 
 
 namespace Westwind.Utilities
@@ -83,6 +84,42 @@ namespace Westwind.Utilities
             return exceptions.ToArray();
         }
 
-	}
+
+        /// <summary>
+        /// Returns the text with a prefix of line numbers
+        /// </summary>
+        /// <param name="text"></param>
+        /// <param name="lineFormat">Line format used to create the line. 0 is the line number, 1 is the text.</param>
+        /// <returns></returns>
+        public static string GetTextWithLineNumbers(string text, string lineFormat = "{0}.  {1}")
+        {
+            if (string.IsNullOrEmpty(text))
+                return text;
+
+            var sb = new StringBuilder();
+            var lines = text.GetLines();
+
+            var width = 2;
+            if (lines.Length > 9999)
+                width = 5;
+            else if (lines.Length > 999)
+                width = 4;
+            else if (lines.Length > 99)
+                width = 3;
+            else if (lines.Length < 10)
+                width = 1;
+
+            lineFormat += "\r\n";
+            for (var index = 1; index <= lines.Length; index++)
+            {
+                var lineNum = index.ToString().PadLeft(width, ' ');
+                sb.AppendFormat(lineFormat, lineNum, lines[index - 1]);
+            }
+
+            return sb.ToString();
+
+        }
+
+    }
 
 }